refactor(audit): move flatpak permission checks to separate module
